+++ adblock.py
@@ -5,45 +5,49 @@ import logging
 import os
 import sys
 import gc
-import psutil
-import time
-import subprocess
 import hashlib
 import json
-from collections import defaultdict, deque
-import asyncio
-import aiodns
-from datetime import datetime, timedelta
+import subprocess
+import time
+from collections import defaultdict
+from datetime import datetime
 from threading import Lock
-import aiohttp
-import smtplib
-from email.mime.text import MIMEText
-import sqlite3
 import socket
-import shelve
-from typing import Dict, List, Optional, Any
-from dataclasses import dataclass
-import pickle
-import zlib
-import backoff
-from writer import safe_save, export_statistics_csv, export_prometheus_metrics
+import aiodns
+
 import aiofiles
+import aiohttp
+import asyncio
+import backoff
+import psutil
 from enum import Enum
-from pybloom_live import ScalableBloomFilter
+
+from caching import (
+    CacheManager,
+    DomainTrie,
+    cleanup_temp_files,
+)
 from config import (
     DB_PATH,
     DEFAULT_CONFIG,
-    HOSTS_HASH_PATH,
     LOG_FORMAT,
-    MAX_DNS_CACHE_SIZE,
     REACHABLE_FILE,
     SCRIPT_DIR,
     TMP_DIR,
-    TRIE_CACHE_PATH,
     UNREACHABLE_FILE,
 )
 from filter_engine import evaluate_lists, parse_domains
+from monitoring import get_system_resources, monitor_resources
+from networking import (
+    is_ipv6_supported,
+    select_best_dns_server,
+    send_email,
+    setup_git,
+    test_domain_batch,
+    upload_to_github,
+)
 from source_loader import load_hosts_sources, load_whitelist_blacklist
+from writer import safe_save, export_statistics_csv, export_prometheus_metrics
 
 
 class SystemMode(Enum):
@@ -110,575 +114,6 @@ STATISTICS = {
 # einer Stelle erfolgen müssen.
 
 logger = logging.getLogger(__name__)
-
-
-class HybridStorage:
-    def __init__(self, db_path: str):
-        logger.debug(f"Initializing HybridStorage with db_path: {db_path}")
-        self.db_path = db_path
-        if os.path.exists(db_path):
-            try:
-                os.remove(db_path)
-                logger.debug(f"Alte Datenbank {db_path} gelöscht")
-            except Exception as e:
-                logger.warning(
-                    f"Fehler beim Löschen der alten Datenbank {db_path}: {e}"
-                )
-        self.db = shelve.open(db_path, protocol=3, writeback=False)
-        logger.debug(f"shelve-Datenbank {db_path} erfolgreich geöffnet")
-        self.ram_storage = {}
-        self.ram_threshold = self.calculate_threshold()
-        self.use_ram = self.should_use_ram()
-        logger.debug(
-            f"HybridStorage initialisiert: RAM-Speicher={self.use_ram}, "
-            f"Pfad={db_path}, Schwellwert={self.ram_threshold/(1024*1024):.2f} MB"
-        )
-
-    def calculate_threshold(self) -> int:
-        try:
-            free_memory = psutil.virtual_memory().available
-            threshold = max(
-                20 * 1024 * 1024, min(512 * 1024 * 1024, int(free_memory * 0.1))
-            )
-            logger.debug(
-                f"Schwellwert berechnet: {threshold/(1024*1024):.2f} MB (freier RAM: {free_memory/(1024*1024):.2f} MB)"
-            )
-            return threshold
-        except Exception as e:
-            logger.warning(
-                f"Fehler bei Schwellwertberechnung: {e}, verwende Fallback: 20 MB"
-            )
-            return 20 * 1024 * 1024
-
-    def update_threshold(self):
-        old_threshold = self.ram_threshold
-        self.ram_threshold = self.calculate_threshold()
-        if old_threshold != self.ram_threshold:
-            logger.debug(
-                f"Schwellwert aktualisiert: {self.ram_threshold/(1024*1024):.2f} MB"
-            )
-        self.use_ram = self.should_use_ram()
-
-    def should_use_ram(self) -> bool:
-        try:
-            free_memory = psutil.virtual_memory().available
-            logger.debug(
-                f"Freier RAM: {free_memory/(1024*1024):.2f} MB, Schwellwert: {self.ram_threshold/(1024*1024):.2f} MB"
-            )
-            return (
-                free_memory > self.ram_threshold
-                or free_memory < 50 * 1024 * 1024
-                or global_mode == SystemMode.EMERGENCY
-            )
-        except Exception as e:
-            logger.warning(
-                f"Fehler beim Überprüfen des freien RAM: {e}, verwende RAM-Speicher"
-            )
-            return True
-
-    def flush_to_disk(self):
-        if self.use_ram:
-            for key, value in self.ram_storage.items():
-                self.db[key] = value
-            self.db.sync()
-            logger.debug("RAM-Daten auf Platte geschrieben")
-
-    def reset_if_corrupt(self):
-        try:
-            if not hasattr(self, "db_path") or not self.db_path:
-                logger.error(
-                    "db_path nicht definiert, kann Datenbank nicht zurücksetzen"
-                )
-                raise ValueError("db_path nicht initialisiert")
-            self.db.close()
-            if os.path.exists(self.db_path):
-                os.remove(self.db_path)
-                logger.warning(
-                    f"Datenbank {self.db_path} zurückgesetzt wegen Korruption"
-                )
-            self.db = shelve.open(self.db_path, protocol=3, writeback=False)
-            self.ram_storage.clear()
-            logger.debug(f"shelve-Datenbank {self.db_path} neu initialisiert")
-        except Exception as e:
-            logger.error(f"Fehler beim Zurücksetzen der Datenbank {self.db_path}: {e}")
-            raise
-
-    def __setitem__(self, key: str, value: Any):
-        try:
-            key = str(key)
-            if self.use_ram:
-                self.ram_storage[key] = value
-            else:
-                self.db[key] = value
-                self.db.sync()
-        except Exception as e:
-            logger.error(
-                f"Kritischer Fehler beim Schreiben in HybridStorage: key={key}, value_type={type(value)}, error={e}"
-            )
-            raise
-
-    def __getitem__(self, key: str) -> Any:
-        key = str(key)
-        if self.use_ram:
-            return self.ram_storage[key]
-        return self.db[key]
-
-    def __contains__(self, key: str) -> bool:
-        key = str(key)
-        if self.use_ram:
-            return key in self.ram_storage
-        return key in self.db
-
-    def __delitem__(self, key: str):
-        key = str(key)
-        if self.use_ram:
-            del self.ram_storage[key]
-        else:
-            del self.db[key]
-            self.db.sync()
-
-    def items(self):
-        if self.use_ram:
-            return self.ram_storage.items()
-        return self.db.items()
-
-    def __len__(self):
-        if self.use_ram:
-            return len(self.ram_storage)
-        return len(self.db)
-
-    def clear(self):
-        self.flush_to_disk()
-        if self.use_ram:
-            self.ram_storage.clear()
-        else:
-            self.db.clear()
-            self.db.sync()
-
-    def close(self):
-        try:
-            self.db.close()
-            logger.debug("HybridStorage-Datenbank geschlossen")
-        except Exception as e:
-            logger.warning(f"Fehler beim Schließen der HybridStorage-Datenbank: {e}")
-
-
-@dataclass
-class TrieNode:
-    children: Dict[str, str]
-    is_end: bool = False
-
-    def __init__(self):
-        self.children = {}
-        self.is_end = False
-
-    def __getstate__(self):
-        return {"children": self.children, "is_end": self.is_end}
-
-    def __setstate__(self, state):
-        self.children = state["children"]
-        self.is_end = state["is_end"]
-
-
-class DomainTrie:
-    def __init__(self, url: str):
-        url_hash = hashlib.md5(url.encode("utf-8")).hexdigest()
-        db_path = os.path.join(TMP_DIR, f"trie_cache_{url_hash}.db")
-        logger.debug(f"Initializing DomainTrie with db_path: {db_path} for URL: {url}")
-        self.storage = HybridStorage(db_path)
-        self.root_key = "root"
-        if CONFIG["use_bloom_filter"] and global_mode != SystemMode.EMERGENCY:
-            self.bloom_filter = ScalableBloomFilter(
-                initial_capacity=CONFIG["bloom_filter_capacity"],
-                error_rate=CONFIG["bloom_filter_error_rate"],
-            )
-        else:
-            self.bloom_filter = None
-        try:
-            root_node = self.storage[self.root_key]
-            if not isinstance(root_node, TrieNode):
-                logger.warning(
-                    f"Root-Knoten ungültig (Typ: {type(root_node)}), initialisiere neu"
-                )
-                self.storage.reset_if_corrupt()
-                self.storage[self.root_key] = TrieNode()
-        except KeyError:
-            logger.debug("Root-Knoten nicht vorhanden, initialisiere neu")
-            self.storage[self.root_key] = TrieNode()
-        logger.debug("DomainTrie initialisiert")
-
-    def insert(self, domain: str):
-        try:
-            if self.bloom_filter and domain in self.bloom_filter:
-                logger.debug(f"Domain {domain} bereits im Bloom-Filter")
-                return
-            logger.debug(f"Inserting domain: {domain}")
-            try:
-                root_node = self.storage[self.root_key]
-                if not isinstance(root_node, TrieNode):
-                    logger.error(
-                        f"Root-Knoten ist kein TrieNode, sondern: {type(root_node)}"
-                    )
-                    self.storage.reset_if_corrupt()
-                    self.storage[self.root_key] = TrieNode()
-            except KeyError:
-                logger.warning("Root-Knoten nicht vorhanden, initialisiere neu")
-                self.storage[self.root_key] = TrieNode()
-            node = self.storage[self.root_key]
-            node_key = self.root_key
-            parts = domain.split(".")[::-1]
-            for i, part in enumerate(parts):
-                logger.debug(f"Processing part {part} at level {i}")
-                if part not in node.children:
-                    new_key = f"{node_key}:{part}:{i}"
-                    node.children[part] = new_key
-                    self.storage[new_key] = TrieNode()
-                    logger.debug(f"Created new node with key {new_key}")
-                node_key = node.children[part]
-                node = self.storage[node_key]
-                logger.debug(f"Moved to node with key {node_key}")
-            node.is_end = True
-            self.storage[node_key] = node
-            if self.bloom_filter:
-                self.bloom_filter.add(domain)
-            logger.debug(f"Domain {domain} successfully inserted")
-        except Exception as e:
-            logger.error(f"Fehler beim Einfügen der Domain {domain} in den Trie: {e}")
-            raise
-
-    def has_parent(self, domain: str) -> bool:
-        try:
-            logger.debug(f"Checking parent for domain: {domain}")
-            parts = domain.split(".")[::-1]
-            node = self.storage[self.root_key]
-            logger.debug(f"Root node retrieved: {node}")
-            node_key = self.root_key
-            for i, part in enumerate(parts):
-                logger.debug(f"Checking part {part} at level {i}")
-                if part not in node.children:
-                    logger.debug(f"Part {part} not found in children")
-                    return False
-                node_key = node.children[part]
-                node = self.storage[node_key]
-                logger.debug(f"Moved to node with key {node_key}")
-                if node.is_end and i < len(parts) - 1:
-                    logger.debug(f"Parent domain found at level {i}")
-                    return True
-            logger.debug(f"No parent domain found for {domain}")
-            return False
-        except Exception as e:
-            logger.error(f"Fehler beim Prüfen der Eltern-Domain für {domain}: {e}")
-            return False
-
-    def flush(self):
-        try:
-            self.storage.clear()
-            self.storage[self.root_key] = TrieNode()
-            logger.debug("DomainTrie Speicher geflusht")
-        except Exception as e:
-            logger.warning(f"Fehler beim Flushen des DomainTrie: {e}")
-
-    def close(self):
-        self.storage.close()
-        logger.debug("DomainTrie geschlossen")
-
-
-def save_trie_cache(trie: DomainTrie, all_domains_hash: str):
-    if not CONFIG["cache_trie"] or global_mode == SystemMode.EMERGENCY:
-        return
-    try:
-        data = {"hash": all_domains_hash, "version": "1.0"}
-        compressed = zlib.compress(pickle.dumps(data))
-        with open(TRIE_CACHE_PATH, "wb") as f:
-            f.write(compressed)
-        logger.info(f"Trie-Cache gespeichert: {TRIE_CACHE_PATH}")
-    except Exception as e:
-        logger.warning(f"Fehler beim Speichern des Trie-Caches: {e}")
-
-
-def load_trie_cache(all_domains_hash: str, url: str) -> Optional[DomainTrie]:
-    if (
-        not CONFIG["cache_trie"]
-        or not os.path.exists(TRIE_CACHE_PATH)
-        or global_mode == SystemMode.EMERGENCY
-    ):
-        logger.debug("Trie-Cache nicht verfügbar oder deaktiviert")
-        return None
-    try:
-        with open(TRIE_CACHE_PATH, "rb") as f:
-            compressed = f.read()
-        data = pickle.loads(zlib.decompress(compressed))
-        if data.get("version") != "1.0":
-            logger.debug("Trie-Cache-Version inkompatibel")
-            return None
-        if data["hash"] == all_domains_hash:
-            STATISTICS["trie_cache_hits"] += 1
-            logger.info("Trie-Cache geladen")
-            return DomainTrie(url)
-        logger.debug("Trie-Cache ungültig (Hash geändert)")
-        return None
-    except Exception as e:
-        logger.warning(f"Fehler beim Laden des Trie-Caches: {e}")
-        return None
-
-
-class CacheManager:
-    def __init__(self, db_path: str, flush_interval: int):
-        self.db_path = db_path
-        self.flush_interval = flush_interval
-        self.domain_cache = HybridStorage(os.path.join(TMP_DIR, "domain_cache.db"))
-        self.list_cache = {}
-        self.last_flush = time.time()
-        self.current_cache_size = self.calculate_dynamic_cache_size()
-        self.init_database()
-        logger.info(
-            f"CacheManager initialisiert: Initiale Cache-Größe={self.current_cache_size}"
-        )
-
-    def calculate_dynamic_cache_size(self) -> int:
-        try:
-            free_memory = psutil.virtual_memory().available
-            estimated_cache_size = int(free_memory * 0.05 / (1024 * 1024) * 100)
-            dynamic_size = max(2, min(100000, estimated_cache_size))
-            if (
-                free_memory
-                < CONFIG["resource_thresholds"]["low_memory_mb"] * 1024 * 1024
-            ):
-                dynamic_size = 2
-                log_once(
-                    logging.WARNING,
-                    "Low-Memory-Modus aktiviert: Cache-Größe auf 2 reduziert",
-                    console=True,
-                )
-            logger.debug(
-                f"Dynamische Cache-Größe berechnet: {dynamic_size} (freier RAM: {free_memory/(1024*1024):.2f} MB)"
-            )
-            return dynamic_size
-        except Exception as e:
-            logger.warning(
-                f"Fehler bei Cache-Größenberechnung: {e}, verwende Fallback: 2"
-            )
-            return 2
-
-    def adjust_cache_size(self):
-        try:
-            new_size = self.calculate_dynamic_cache_size()
-            if new_size != self.current_cache_size:
-                self.current_cache_size = new_size
-                logger.info(f"Cache-Größe angepasst: {self.current_cache_size}")
-                if (
-                    len(self.domain_cache.ram_storage) > self.current_cache_size
-                    and self.domain_cache.use_ram
-                ):
-                    sorted_items = sorted(
-                        self.domain_cache.ram_storage.items(),
-                        key=lambda x: x[1]["checked_at"],
-                    )
-                    self.domain_cache.ram_storage = dict(
-                        sorted_items[-self.current_cache_size :]
-                    )
-                    logger.debug(
-                        f"Domain-Cache beschnitten auf {self.current_cache_size} Einträge"
-                    )
-        except Exception as e:
-            logger.warning(f"Fehler beim Anpassen der Cache-Größe: {e}")
-
-    def init_database(self):
-        try:
-            os.makedirs(TMP_DIR, exist_ok=True)
-            conn = sqlite3.connect(self.db_path, timeout=30)
-            c = conn.cursor()
-            c.execute(
-                """
-                CREATE TABLE IF NOT EXISTS list_cache (
-                    url TEXT PRIMARY KEY,
-                    md5 TEXT,
-                    last_checked TEXT
-                )
-            """
-            )
-            conn.commit()
-            conn.close()
-            logger.debug("SQLite-Datenbank initialisiert")
-        except Exception as e:
-            logger.error(f"Fehler beim Initialisieren der SQLite-Datenbank: {e}")
-            raise
-
-    async def flush_cache_periodically(self):
-        while True:
-            try:
-                await asyncio.sleep(5)
-                memory = psutil.Process().memory_info().rss / (1024 * 1024)
-                free_memory = psutil.virtual_memory().available / (1024 * 1024)
-                cpu_usage = psutil.cpu_percent(interval=0.1)
-                if free_memory < 50 or psutil.virtual_memory().percent > 90:
-                    log_once(
-                        logging.WARNING,
-                        f"Kritischer Speicherstand: {free_memory:.2f} MB frei, "
-                        f"Auslastung: {psutil.virtual_memory().percent}%, reduziere Cache",
-                        console=True,
-                    )
-                    self.current_cache_size = max(2, self.current_cache_size // 2)
-                    self.adjust_cache_size()
-                    self.domain_cache.use_ram = True
-                elif free_memory > 512 and psutil.virtual_memory().percent < 50:
-                    self.adjust_cache_size()
-                    self.domain_cache.use_ram = True
-                if cpu_usage > 90:
-                    log_once(
-                        logging.WARNING,
-                        f"Hohe CPU-Auslastung: {cpu_usage}%, reduziere parallele Aufgaben",
-                        console=True,
-                    )
-                logger.debug(
-                    f"Speicherverbrauch: {memory:.2f} MB, CPU: {cpu_usage}%, "
-                    f"Cache-Größe: {self.current_cache_size}, RAM-Speicher: {self.domain_cache.use_ram}"
-                )
-                if time.time() - self.last_flush > self.flush_interval:
-                    async with cache_flush_lock:
-                        self.save_domain_cache()
-                        self.last_flush = time.time()
-            except asyncio.CancelledError:
-                async with cache_flush_lock:
-                    self.save_domain_cache()
-                break
-            except Exception as e:
-                logger.warning(f"Fehler im Cache-Flush-Task: {e}")
-
-    def load_domain_cache(self):
-        try:
-            return self.domain_cache
-        except Exception as e:
-            logger.warning(f"Fehler beim Zugriff auf Domain-Cache: {e}")
-            return {}
-
-    def save_domain_cache(self):
-        try:
-            if (
-                len(self.domain_cache.ram_storage) > self.current_cache_size
-                and self.domain_cache.use_ram
-            ):
-                sorted_items = sorted(
-                    self.domain_cache.ram_storage.items(),
-                    key=lambda x: x[1]["checked_at"],
-                )
-                self.domain_cache.ram_storage = dict(
-                    sorted_items[-self.current_cache_size :]
-                )
-                logger.debug(
-                    f"Domain-Cache beschnitten auf {self.current_cache_size} Einträge"
-                )
-            STATISTICS["cache_flushes"] += 1
-            logger.debug(
-                f"Domain-Cache geflusht (Flush #{STATISTICS['cache_flushes']})"
-            )
-        except Exception as e:
-            logger.warning(f"Fehler beim Speichern des Domain-Caches: {e}")
-
-    def save_domain(self, domain: str, reachable: bool, source_url: str):
-        try:
-            self.domain_cache[domain] = {
-                "reachable": reachable,
-                "checked_at": datetime.now().isoformat(),
-                "source": source_url,
-            }
-            if (
-                len(self.domain_cache.ram_storage) > self.current_cache_size
-                and self.domain_cache.use_ram
-            ):
-                oldest_domain = min(
-                    self.domain_cache.ram_storage,
-                    key=lambda k: self.domain_cache.ram_storage[k]["checked_at"],
-                )
-                del self.domain_cache[oldest_domain]
-                logger.debug(f"Älteste Domain {oldest_domain} aus Cache entfernt")
-        except Exception as e:
-            logger.warning(f"Fehler beim Speichern der Domain {domain}: {e}")
-
-    def load_list_cache(self) -> Dict[str, Dict]:
-        try:
-            conn = sqlite3.connect(self.db_path, timeout=30)
-            c = conn.cursor()
-            c.execute("SELECT url, md5, last_checked FROM list_cache")
-            self.list_cache = {
-                row[0]: {"md5": row[1], "last_checked": row[2]} for row in c.fetchall()
-            }
-            conn.close()
-            logger.debug("List-Cache geladen")
-            return self.list_cache
-        except Exception as e:
-            logger.warning(f"Fehler beim Zugriff auf List-Cache: {e}")
-            return {}
-
-    def save_list_cache(self, cache: Dict[str, Dict]):
-        try:
-            conn = sqlite3.connect(self.db_path, timeout=30)
-            c = conn.cursor()
-            c.execute("DELETE FROM list_cache")
-            for url, data in cache.items():
-                c.execute(
-                    """
-                    INSERT INTO list_cache (url, md5, last_checked)
-                    VALUES (?, ?, ?)
-                """,
-                    (url, data["md5"], data["last_checked"]),
-                )
-            conn.commit()
-            conn.close()
-            logger.debug("List-Cache gespeichert")
-        except Exception as e:
-            logger.warning(f"Fehler beim Speichern des List-Caches: {e}, fahre fort")
-
-
-def cleanup_temp_files(cache_manager: CacheManager):
-    try:
-        list_cache = cache_manager.load_list_cache()
-        valid_urls = set(list_cache.keys())
-        expiry = datetime.now() - timedelta(days=CONFIG["domain_cache_validity_days"])
-        for file in os.listdir(TMP_DIR):
-            if (
-                file.endswith(".tmp")
-                or file.endswith(".filtered")
-                or file.startswith("trie_cache_")
-            ):
-                file_path = os.path.join(TMP_DIR, file)
-                url = (
-                    file.replace(".tmp", "")
-                    .replace(".filtered", "")
-                    .replace("trie_cache_", "")
-                    .replace("__", "/")
-                    .replace("_", "://")
-                )
-                if file.startswith("trie_cache_"):
-                    if url in list_cache:
-                        last_checked = datetime.fromisoformat(
-                            list_cache[url]["last_checked"]
-                        )
-                        if last_checked < expiry:
-                            try:
-                                os.remove(file_path)
-                                logger.info(
-                                    f"Veraltete Trie-Cache-Datei gelöscht: {file}"
-                                )
-                            except Exception as e:
-                                logger.error(f"Fehler beim Löschen von {file}: {e}")
-                        else:
-                            logger.debug(f"Trie-Cache-Datei behalten: {file}")
-                    else:
-                        try:
-                            os.remove(file_path)
-                            logger.info(f"Obsolete Trie-Cache-Datei gelöscht: {file}")
-                        except Exception as e:
-                            logger.error(f"Fehler beim Löschen von {file}: {e}")
-                elif url not in valid_urls:
-                    try:
-                        os.remove(file_path)
-                        logger.info(f"Obsolete Datei gelöscht: {file}")
-                    except Exception as e:
-                        logger.error(f"Fehler beim Löschen von {file}: {e}")
-        logger.debug("Temporäre Dateien bereinigt")
-    except Exception as e:
-        logger.error(f"Fehler beim Bereinigen temporärer Dateien: {e}")
 
 
 def load_config():
@@ -871,53 +306,6 @@ def calculate_md5(content: str) -> str:
         return ""
 
 
-def send_email(subject: str, body: str):
-    if not CONFIG.get("send_email", False) or global_mode == SystemMode.EMERGENCY:
-        return
-    try:
-        msg = MIMEText(body)
-        msg["Subject"] = subject
-        msg["From"] = CONFIG["email_sender"]
-        msg["To"] = CONFIG["email_recipient"]
-
-        if CONFIG["use_smtp"]:
-            with smtplib.SMTP(CONFIG["smtp_server"], CONFIG["smtp_port"]) as server:
-                server.starttls()
-                server.login(CONFIG["smtp_user"], CONFIG["smtp_password"])
-                server.send_message(msg)
-            logger.info("E-Mail-Benachrichtigung über SMTP gesendet")
-        else:
-            sendmail_cmd = ["/usr/sbin/sendmail", "-t", "-oi"]
-            process = subprocess.Popen(sendmail_cmd, stdin=subprocess.PIPE)
-            process.communicate(msg.as_string().encode("utf-8"))
-            if process.returncode == 0:
-                logger.info("E-Mail-Benachrichtigung über lokales Postfix gesendet")
-            else:
-                raise Exception(f"sendmail failed with exit code {process.returncode}")
-    except Exception as e:
-        logger.error(f"Fehler beim Senden der E-Mail: {e}")
-
-
-async def is_ipv6_supported() -> bool:
-    try:
-        ipv6_servers = [server for server in CONFIG["dns_servers"] if ":" in server]
-        if not ipv6_servers:
-            logger.debug("Keine IPv6-DNS-Server in der Konfiguration gefunden")
-            return False
-        resolver = aiodns.DNSResolver(nameservers=ipv6_servers, timeout=2)
-        await resolver.query("example.com", "AAAA")
-        logger.debug(
-            "IPv6 wird unterstützt (erfolgreicher AAAA-Lookup über IPv6-DNS-Server)"
-        )
-        return True
-    except (aiodns.error.DNSError, asyncio.TimeoutError) as e:
-        logger.debug(f"IPv6 wird nicht unterstützt: {e}")
-        return False
-    except Exception as e:
-        logger.warning(f"Unbekannter Fehler beim Testen von IPv6: {e}")
-        return False
-
-
 def initialize_directories_and_files():
     try:
         os.makedirs(TMP_DIR, exist_ok=True)
@@ -951,554 +339,6 @@ def initialize_directories_and_files():
 # Die Funktionen zum Laden der Quelllisten sowie von White- und Blacklist sind
 # bereits im Modul ``source_loader`` implementiert und werden hier nur noch
 # importiert. Dadurch vermeiden wir doppelte Logik innerhalb des Projekts.
-
-
-async def select_best_dns_server(
-    dns_servers: List[str], timeout: float = 5.0
-) -> List[str]:
-    async def test_server(server: str) -> tuple[str, float]:
-        try:
-            resolver = aiodns.DNSResolver(nameservers=[server], timeout=timeout)
-            start = time.time()
-            await resolver.query("example.com", "A")
-            latency = time.time() - start
-            logger.debug(f"DNS-Server {server} Latenz: {latency:.2f}s")
-            return server, latency
-        except Exception as e:
-            logger.debug(f"DNS-Server {server} nicht erreichbar: {e}")
-            return server, float("inf")
-
-    try:
-        tasks = [test_server(server) for server in dns_servers]
-        results = await asyncio.gather(*tasks)
-        sorted_servers = [
-            server
-            for server, latency in sorted(results, key=lambda x: x[1])
-            if latency != float("inf")
-        ]
-        if not sorted_servers:
-            logger.warning("Kein DNS-Server erreichbar, verwende Standardserver")
-            return dns_servers
-        logger.debug(f"Ausgewählte DNS-Server: {sorted_servers}")
-        return sorted_servers
-    except Exception as e:
-        logger.error(f"Fehler beim Auswählen der DNS-Server: {e}")
-        return dns_servers
-
-
-async def monitor_resources(cache_manager: CacheManager):
-    global global_mode
-    logger.info("Starte Ressourcenüberwachung...")
-    resource_state = {
-        "low_memory": False,
-        "high_cpu": False,
-        "high_latency": False,
-        "mode": global_mode,
-    }
-    consecutive_violations = {"low_memory": 0, "high_cpu": 0, "high_latency": 0}
-    history = {
-        "free_memory_mb": deque(
-            maxlen=CONFIG["resource_thresholds"]["moving_average_window"]
-        ),
-        "cpu_usage_percent": deque(
-            maxlen=CONFIG["resource_thresholds"]["moving_average_window"]
-        ),
-        "latency_s": deque(
-            maxlen=CONFIG["resource_thresholds"]["moving_average_window"]
-        ),
-    }
-    last_status_log = time.time()
-    first_run = True
-
-    def calculate_moving_average(values: deque) -> float:
-        return sum(values) / len(values) if values else 0.0
-
-    def print_system_status():
-        log_once(
-            logging.INFO,
-            f"Systemzustandsübersicht:\n"
-            f"- Aktueller Modus: {resource_state['mode'].value}\n"
-            f"- Freier RAM: {avg_memory:.2f} MB\n"
-            f"- CPU-Last: {avg_cpu:.1f}%\n"
-            f"- DNS-Latenz: {avg_latency:.2f}s\n"
-            f"- Angepasste Batch-Größe: {batch_size} Domains",
-            console=True,
-        )
-
-    while True:
-        try:
-            free_memory = psutil.virtual_memory().available / (1024 * 1024)
-            cpu_usage = psutil.cpu_percent(interval=0.1)
-            latency = await check_network_latency()
-            max_jobs, batch_size, max_concurrent_dns = get_system_resources()
-
-            history["free_memory_mb"].append(free_memory)
-            history["cpu_usage_percent"].append(cpu_usage)
-            history["latency_s"].append(latency if latency != float("inf") else 5.0)
-
-            avg_memory = calculate_moving_average(history["free_memory_mb"])
-            avg_cpu = calculate_moving_average(history["cpu_usage_percent"])
-            avg_latency = calculate_moving_average(history["latency_s"])
-
-            new_mode = resource_state["mode"]
-            if avg_memory < CONFIG["resource_thresholds"]["emergency_memory_mb"]:
-                new_mode = SystemMode.EMERGENCY
-            elif (
-                avg_memory < CONFIG["resource_thresholds"]["low_memory_mb"]
-                or avg_cpu > CONFIG["resource_thresholds"]["high_cpu_percent"]
-                or avg_latency > CONFIG["resource_thresholds"]["high_latency_s"]
-            ):
-                new_mode = SystemMode.LOW_MEMORY
-            else:
-                new_mode = SystemMode.NORMAL
-
-            if new_mode != resource_state["mode"]:
-                log_once(
-                    logging.INFO,
-                    f"Betriebsmodus gewechselt: {resource_state['mode'].value} → {new_mode.value}",
-                    console=True,
-                )
-                resource_state["mode"] = new_mode
-                global_mode = new_mode
-
-            if avg_memory < CONFIG["resource_thresholds"]["low_memory_mb"]:
-                consecutive_violations["low_memory"] += 1
-                if (
-                    consecutive_violations["low_memory"]
-                    >= CONFIG["resource_thresholds"]["consecutive_violations"]
-                    and not resource_state["low_memory"]
-                ):
-                    log_once(
-                        logging.WARNING,
-                        f"Wenig RAM verfügbar: Durchschnitt {avg_memory:.2f} MB, "
-                        f"Batch-Größe: {batch_size}, DNS-Anfragen: {max_concurrent_dns}",
-                        console=True,
-                    )
-                    resource_state["low_memory"] = True
-            else:
-                consecutive_violations["low_memory"] = 0
-                if resource_state["low_memory"]:
-                    log_once(
-                        logging.INFO,
-                        f"RAM wieder ausreichend verfügbar: Durchschnitt {avg_memory:.2f} MB frei",
-                        console=True,
-                    )
-                    resource_state["low_memory"] = False
-
-            if avg_cpu > CONFIG["resource_thresholds"]["high_cpu_percent"]:
-                consecutive_violations["high_cpu"] += 1
-                if (
-                    consecutive_violations["high_cpu"]
-                    >= CONFIG["resource_thresholds"]["consecutive_violations"]
-                    and not resource_state["high_cpu"]
-                ):
-                    log_once(
-                        logging.WARNING,
-                        f"Hohe CPU-Auslastung erkannt: Durchschnitt {avg_cpu:.1f}%, reduziere parallele Jobs",
-                        console=True,
-                    )
-                    resource_state["high_cpu"] = True
-            else:
-                consecutive_violations["high_cpu"] = 0
-                if resource_state["high_cpu"]:
-                    log_once(
-                        logging.INFO,
-                        f"CPU-Auslastung wieder normal: Durchschnitt {avg_cpu:.1f}%",
-                        console=True,
-                    )
-                    resource_state["high_cpu"] = False
-
-            if avg_latency > CONFIG["resource_thresholds"]["high_latency_s"]:
-                consecutive_violations["high_latency"] += 1
-                if (
-                    consecutive_violations["high_latency"]
-                    >= CONFIG["resource_thresholds"]["consecutive_violations"]
-                    and not resource_state["high_latency"]
-                ):
-                    log_once(
-                        logging.WARNING,
-                        f"Hohe Netzwerklatenz erkannt: Durchschnitt {avg_latency:.2f}s, DNS-Anfragen reduziert",
-                        console=True,
-                    )
-                    resource_state["high_latency"] = True
-            else:
-                consecutive_violations["high_latency"] = 0
-                if resource_state["high_latency"]:
-                    log_once(
-                        logging.INFO,
-                        f"Netzwerklatenz wieder normal: Durchschnitt {avg_latency:.2f}s",
-                        console=True,
-                    )
-                    resource_state["high_latency"] = False
-
-            cache_manager.adjust_cache_size()
-            cache_manager.domain_cache.update_threshold()
-
-            if first_run or time.time() - last_status_log >= 300:
-                print_system_status()
-                logger.info(
-                    f"Systemstatus (Modus: {resource_state['mode'].value}): "
-                    f"RAM-Durchschnitt={avg_memory:.2f} MB, "
-                    f"CPU-Durchschnitt={avg_cpu:.1f}%, Latenz-Durchschnitt={avg_latency:.2f}s, "
-                    f"Cachegröße={len(cache_manager.domain_cache.ram_storage)}, "
-                    f"Jobs={max_jobs}, DNS-Anfragen={max_concurrent_dns}"
-                )
-                last_status_log = time.time()
-                first_run = False
-
-            await asyncio.sleep(5)
-
-        except Exception as e:
-            logger.warning(f"Fehler bei Ressourcenüberwachung: {e}")
-            await asyncio.sleep(5)
-
-
-async def check_network_latency() -> float:
-    try:
-        resolver = aiodns.DNSResolver(
-            nameservers=[CONFIG["dns_servers"][0]], timeout=5.0
-        )
-        start = time.time()
-        await resolver.query("example.com", "A")
-        latency = time.time() - start
-        logger.debug(f"Netzwerklatenz zu {CONFIG['dns_servers'][0]}: {latency:.2f}s")
-        return latency
-    except aiodns.error.DNSError as e:
-        logger.warning(f"DNS-Fehler bei Latenzprüfung: {e}")
-        return float("inf")
-    except Exception as e:
-        logger.warning(f"Unbekannter Fehler bei Latenzprüfung: {e}")
-        return float("inf")
-
-
-def get_system_resources() -> tuple[int, int, int]:
-    try:
-        cpu_load = psutil.cpu_percent(interval=0.1) / 100
-        cpu_cores = psutil.cpu_count(logical=True) or 1
-        free_memory = psutil.virtual_memory().available
-
-        if global_mode == SystemMode.EMERGENCY:
-            max_jobs = 1
-            batch_size = 5
-            max_concurrent_dns = 5
-            log_once(
-                logging.WARNING,
-                "Emergency-Mode: Batch-Größe=5, Jobs=1, DNS-Anfragen=5",
-                console=True,
-            )
-        elif global_mode == SystemMode.LOW_MEMORY:
-            max_jobs = max(1, int(cpu_cores / (cpu_load + 0.1)) // 4)
-            batch_size = max(5, min(20, int(free_memory / (1000 * 1024))))
-            max_concurrent_dns = max(5, min(8, int(free_memory / (2048 * 1024))))
-            log_once(
-                logging.WARNING,
-                f"Low-Memory-Modus: Batch-Größe={batch_size}, Jobs={max_jobs}, DNS-Anfragen={max_concurrent_dns}",
-                console=True,
-            )
-        else:
-            max_jobs = max(
-                1,
-                int(cpu_cores / (cpu_load + 0.1)) // 2,
-            )
-            batch_size = max(
-                10,
-                min(50, int(free_memory / (500 * 1024))),
-            )
-            max_concurrent_dns = max(
-                5,
-                min(20, int(free_memory / (1024 * 1024))),
-            )
-
-        logger.debug(
-            f"CPU-Last: {cpu_load:.2f}, Kerne: {cpu_cores}, "
-            f"Speicher: {free_memory/(1024*1024):.2f} MB, Modus: {global_mode.value}"
-        )
-        logger.debug(
-            f"Empfohlene Jobs: {max_jobs}, Batch-Größe: {batch_size}, DNS-Anfragen: {max_concurrent_dns}"
-        )
-        return max_jobs, batch_size, max_concurrent_dns
-    except Exception as e:
-        logger.error(f"Fehler bei Ressourcenermittlung: {e}, verwende Fallback")
-        return 1, 5, 5
-
-
-async def test_dns_entry_async(
-    domain: str, resolver, record_type: str = "A", max_concurrent: int = 5
-) -> bool:
-    async def query_with_backoff(
-        domain: str, record: str, resolver, attempt: int
-    ) -> bool:
-        try:
-            result = await resolver.query(domain, record)
-            return bool(result)
-        except aiodns.error.DNSError as e:
-            if "NXDOMAIN" in str(e):
-                logger.debug(f"Domain {domain} nicht existent (NXDOMAIN, {record})")
-                return False
-            logger.debug(f"Fehler bei {domain} ({record}, Versuch {attempt + 1}): {e}")
-            return False
-        except Exception as e:
-            logger.debug(
-                f"Unbekannter Fehler bei {domain} ({record}, Versuch {attempt + 1}): {e}"
-            )
-            return False
-
-    try:
-        with dns_cache_lock:
-            if domain in DNS_CACHE:
-                logger.debug(f"Domain {domain} aus Cache: {DNS_CACHE[domain]}")
-                return DNS_CACHE[domain]
-            if len(DNS_CACHE) >= MAX_DNS_CACHE_SIZE:
-                oldest_domain = next(iter(DNS_CACHE))
-                del DNS_CACHE[oldest_domain]
-                logger.debug(f"Älteste Domain {oldest_domain} aus DNS-Cache entfernt")
-
-        semaphore = asyncio.Semaphore(max_concurrent)
-        async with semaphore:
-            record_types = (
-                [record_type, "AAAA"] if record_type == "A" else [record_type]
-            )
-            for record in record_types:
-                for attempt in range(CONFIG["max_retries"]):
-                    reachable = await query_with_backoff(
-                        domain, record, resolver, attempt
-                    )
-                    if reachable:
-                        with dns_cache_lock:
-                            DNS_CACHE[domain] = True
-                        logger.debug(f"Domain {domain} erreichbar ({record})")
-                        return True
-                    delay = (
-                        CONFIG["retry_delay"] * (2**attempt)
-                        if CONFIG["dns_retry_strategy"] == "exponential"
-                        else CONFIG["retry_delay"]
-                    )
-                    await asyncio.sleep(delay)
-            with dns_cache_lock:
-                DNS_CACHE[domain] = False
-            logger.debug(f"Domain {domain} nicht erreichbar")
-            return False
-    except Exception as e:
-        logger.error(f"Kritischer Fehler beim Testen der DNS-Domain {domain}: {e}")
-        return False
-
-
-async def test_single_domain_async(
-    domain: str,
-    url: str,
-    resolver,
-    cache_manager: CacheManager,
-    whitelist: set[str],
-    blacklist: set[str],
-    max_concurrent: int = 5,
-) -> bool:
-    try:
-        if domain in whitelist:
-            logger.debug(f"Domain {domain} in Whitelist, überspringe Prüfung")
-            return False
-        if domain in blacklist:
-            logger.debug(f"Domain {domain} in Blacklist, blockiere immer")
-            return True
-        if not CONFIG["always_check_all_domains"]:
-            cache = cache_manager.load_domain_cache()
-            if domain in cache:
-                entry = cache[domain]
-                last_checked = datetime.fromisoformat(entry["checked_at"])
-                if datetime.now() - last_checked < timedelta(
-                    days=CONFIG["domain_cache_validity_days"]
-                ):
-                    logger.debug(f"Domain {domain} aus Cache: {entry['reachable']}")
-                    STATISTICS["cache_hits"] += 1
-                    STATISTICS["list_stats"][url][
-                        "reachable" if entry["reachable"] else "unreachable"
-                    ] += 1
-                    return entry["reachable"]
-        reachable = await test_dns_entry_async(
-            domain, resolver, max_concurrent=max_concurrent
-        )
-        cache_manager.save_domain(domain, reachable, url)
-        STATISTICS["list_stats"][url]["reachable" if reachable else "unreachable"] += 1
-        return reachable
-    except Exception as e:
-        logger.error(f"Fehler beim Testen der Domain {domain} für URL {url}: {e}")
-        return False
-
-
-async def test_domain_batch(
-    domains: List[str],
-    url: str,
-    resolver,
-    cache_manager: CacheManager,
-    whitelist: set[str],
-    blacklist: set[str],
-    max_concurrent: int = 5,
-) -> List[tuple[str, bool]]:
-    try:
-        free_memory = psutil.virtual_memory().available
-        cache_manager.domain_cache.update_threshold()
-        _, batch_size, max_concurrent = get_system_resources()
-        logger.debug(
-            f"Batch-Größe für DNS-Validierung: {batch_size}, "
-            f"Max. DNS-Anfragen: {max_concurrent}, "
-            f"Freier RAM: {free_memory/(1024*1024):.2f} MB"
-        )
-        results = []
-        tasks = [
-            test_single_domain_async(
-                domain,
-                url,
-                resolver,
-                cache_manager,
-                whitelist,
-                blacklist,
-                max_concurrent,
-            )
-            for domain in domains
-        ]
-        results = await asyncio.gather(*tasks, return_exceptions=True)
-        gc.collect()
-        return [
-            (domain, result)
-            for domain, result in zip(domains, results)
-            if not isinstance(result, Exception)
-        ]
-    except Exception as e:
-        logger.error(f"Fehler beim Testen des Domain-Batches für URL {url}: {e}")
-        return []
-
-
-def setup_git() -> bool:
-    try:
-        # Prüfe, ob git installiert ist
-        subprocess.run(["git", "--version"], check=True, capture_output=True, text=True)
-        logger.debug("Git ist installiert")
-
-        # Prüfe, ob das Repository existiert
-        git_dir = os.path.join(SCRIPT_DIR, ".git")
-        if not os.path.exists(git_dir):
-            logger.error(
-                "Git-Repository nicht gefunden. Bitte initialisieren Sie das Repository manuell."
-            )
-            return False
-
-        # Prüfe, ob das Remote-Repository korrekt konfiguriert ist
-        result = subprocess.run(
-            ["git", "remote", "-v"], capture_output=True, text=True, cwd=SCRIPT_DIR
-        )
-        if CONFIG["github_repo"] not in result.stdout:
-            logger.debug(f"Füge Remote-Repository hinzu: {CONFIG['github_repo']}")
-            subprocess.run(
-                ["git", "remote", "add", "origin", CONFIG["github_repo"]],
-                check=True,
-                cwd=SCRIPT_DIR,
-            )
-        else:
-            logger.debug("Remote-Repository bereits konfiguriert")
-
-        # Prüfe, ob der Branch existiert
-        branch_result = subprocess.run(
-            ["git", "branch", "--list", CONFIG["github_branch"]],
-            capture_output=True,
-            text=True,
-            cwd=SCRIPT_DIR,
-        )
-        if CONFIG["github_branch"] not in branch_result.stdout:
-            logger.debug(f"Erstelle neuen Branch: {CONFIG['github_branch']}")
-            subprocess.run(
-                ["git", "checkout", "-b", CONFIG["github_branch"]],
-                check=True,
-                cwd=SCRIPT_DIR,
-            )
-        else:
-            logger.debug(f"Wechsel zu Branch: {CONFIG['github_branch']}")
-            subprocess.run(
-                ["git", "checkout", CONFIG["github_branch"]], check=True, cwd=SCRIPT_DIR
-            )
-
-        # Teste die SSH-Verbindung zu GitHub
-        result = subprocess.run(
-            ["ssh", "-T", "git@github.com"], capture_output=True, text=True
-        )
-        if "successfully authenticated" in result.stderr or result.returncode == 0:
-            logger.debug("SSH-Verbindung zu GitHub erfolgreich")
-            return True
-        else:
-            logger.error(f"SSH-Verbindung zu GitHub fehlgeschlagen: {result.stderr}")
-            return False
-
-    except subprocess.CalledProcessError as e:
-        logger.warning(f"Fehler bei der Git-Konfiguration: {e}")
-        return False
-    except Exception as e:
-        logger.warning(f"Unbekannter Fehler bei der Git-Konfiguration: {e}")
-        return False
-
-
-def upload_to_github():
-    if not CONFIG["github_upload"] or global_mode == SystemMode.EMERGENCY:
-        logger.info("GitHub-Upload deaktiviert")
-        return
-    hosts_file = os.path.join(SCRIPT_DIR, CONFIG["hosts_file"])
-    try:
-        if not os.path.exists(hosts_file):
-            logger.warning("hosts.txt fehlt, überspringe Git-Upload")
-            return
-        new_hash = calculate_md5(open(hosts_file, "r", encoding="utf-8").read())
-        old_hash = ""
-        if os.path.exists(HOSTS_HASH_PATH):
-            with open(HOSTS_HASH_PATH, "r", encoding="utf-8") as f:
-                old_hash = f.read().strip()
-        if new_hash == old_hash:
-            logger.info("Keine Änderungen an hosts.txt, überspringe Git-Upload")
-            return
-        result = subprocess.run(
-            ["git", "status", "--porcelain", CONFIG["hosts_file"]],
-            capture_output=True,
-            text=True,
-            cwd=SCRIPT_DIR,
-        )
-        if not result.stdout.strip():
-            logger.info("Keine Änderungen an hosts.txt erkannt, überspringe Git-Upload")
-            return
-        subprocess.run(
-            ["git", "add", CONFIG["hosts_file"]],
-            check=True,
-            capture_output=True,
-            text=True,
-            cwd=SCRIPT_DIR,
-        )
-        commit_msg = (
-            f"Update hosts.txt: {STATISTICS['reachable_domains']} Domains, "
-            f"{STATISTICS['unique_domains']} einzigartig"
-        )
-        result = subprocess.run(
-            ["git", "commit", "-m", commit_msg],
-            check=True,
-            capture_output=True,
-            text=True,
-            cwd=SCRIPT_DIR,
-        )
-        logger.debug(f"Git commit Ausgabe: {result.stdout}")
-        result = subprocess.run(
-            ["git", "push", "origin", CONFIG["github_branch"]],
-            check=True,
-            capture_output=True,
-            text=True,
-            cwd=SCRIPT_DIR,
-        )
-        logger.debug(f"Git push Ausgabe: {result.stdout}")
-        logger.info("Hosts-Datei erfolgreich auf GitHub hochgeladen")
-        with open(HOSTS_HASH_PATH, "w", encoding="utf-8") as f:
-            f.write(new_hash)
-    except subprocess.CalledProcessError as e:
-        logger.warning(f"Fehler beim Hochladen auf GitHub: {e}, stderr: {e.stderr}")
-        send_email(
-            "Warnung im AdBlock-Skript",
-            f"Git-Upload fehlgeschlagen: {e}\nStderr: {e.stderr}",
-        )
-    except Exception as e:
-        logger.warning(f"Unbekannter Fehler beim Hochladen auf GitHub: {e}")
-        send_email("Warnung im AdBlock-Skript", f"Git-Upload fehlgeschlagen: {e}")
 
 
 # =============================================================================
@@ -1720,7 +560,9 @@ async def main():
 
         logger.debug("Starte Cache-Flush- und Ressourcenüberwachungs-Tasks...")
         cache_flush_task = asyncio.create_task(cache_manager.flush_cache_periodically())
-        resource_monitor_task = asyncio.create_task(monitor_resources(cache_manager))
+        resource_monitor_task = asyncio.create_task(
+            monitor_resources(cache_manager, CONFIG)
+        )
         logger.debug("Cache-Flush- und Ressourcenüberwachungs-Tasks gestartet")
 
         if CONFIG["github_upload"]:
@@ -1739,6 +581,7 @@ async def main():
                 send_email(
                     "Fehler im AdBlock-Skript",
                     "Keine Quell-URLs in hosts_sources.conf gefunden",
+                    CONFIG,
                 )
             raise ValueError("Keine Quell-URLs gefunden")
         logger.debug(f"Geladene Quell-URLs: {len(sources)}")
@@ -1794,7 +637,9 @@ async def main():
             logger.warning("Keine gültigen Domains gefunden")
             if global_mode != SystemMode.EMERGENCY:
                 send_email(
-                    "Fehler im AdBlock-Skript", "Keine gültigen Domains gefunden"
+                    "Fehler im AdBlock-Skript",
+                    "Keine gültigen Domains gefunden",
+                    CONFIG,
                 )
             processed_urls = []
         STATISTICS["total_domains"] = sum(
@@ -1998,7 +843,7 @@ Empfehlungen:
         logger.info("Zusammenfassung erfolgreich erstellt")
         logger.info(summary)
         if CONFIG.get("send_email", False) and global_mode != SystemMode.EMERGENCY:
-            send_email("AdBlock-Skript Bericht", summary)
+            send_email("AdBlock-Skript Bericht", summary, CONFIG)
         try:
             subprocess.run(["systemctl", "restart", "dnsmasq"], check=True)
             logger.info("DNSMasq erfolgreich neu gestartet")
@@ -2015,13 +860,16 @@ Empfehlungen:
                     send_email(
                         "Fehler im AdBlock-Skript",
                         f"DNSMasq-Neustart fehlgeschlagen: {e}",
+                        CONFIG,
                     )
         logger.info("Skript erfolgreich abgeschlossen")
     except Exception as e:
         logger.error(f"Kritischer Fehler in der Hauptfunktion: {e}")
         if global_mode != SystemMode.EMERGENCY:
             send_email(
-                "Kritischer Fehler im AdBlock-Skript", f"Skript fehlgeschlagen: {e}"
+                "Kritischer Fehler im AdBlock-Skript",
+                f"Skript fehlgeschlagen: {e}",
+                CONFIG,
             )
         sys.exit(1)
     finally:
@@ -2055,12 +903,15 @@ if __name__ == "__main__":
             send_email(
                 "Kritischer Fehler im AdBlock-Skript",
                 "Skript abgebrochen: Nicht genügend Speicher verfügbar",
+                CONFIG,
             )
         sys.exit(1)
     except Exception as e:
         logger.error(f"Kritischer Fehler beim Start des Skripts: {e}")
         if global_mode != SystemMode.EMERGENCY:
             send_email(
-                "Kritischer Fehler im AdBlock-Skript", f"Skript fehlgeschlagen: {e}"
+                "Kritischer Fehler im AdBlock-Skript",
+                f"Skript fehlgeschlagen: {e}",
+                CONFIG,
             )
         sys.exit(1)
